requirements.txt
psycopg2
pandas
sqlalchemy
scholarly
alembic
colorlog
python-dotenv